requirements.txt
flask==2.2.3
gunicorn==20.1.0
pytesseract==0.3.10
Pillow==9.5.0
PyMuPDF==1.22.5
google-generativeai==0.3.1
Werkzeug==2.2.3